refactor(payloadstore): remove config and serializer files, refactor payloadstore to include them
